refactor: Refactored the Logger struct and added a new method to convert log levels
This commit refactors the Logger struct by moving its definition inside a type block. It also adds a new method, ConvertLogLevel, which converts log levels from an int to a string or from a number as a string to a log level name. The LogLevels map is used for this conversion.