refactor: Change getCtx and getLogPrefix functions to accept pointer
This commit changes the`getCtx` and `getLogPrefix` functions in `logger.go` to accept a pointer to string instead of a regular string. This allows for `nil` values to be passed as context, which is useful when calling the `Debug`, `Info`, `Warn`, `Error`, and `Fatal` functions without specifying a context. The commit also adds a new function `GetCallingFunctionName` that returns the name of the calling function. Finally, this commit updates all calls to `Debug`, `Info`, `Warn`, `Error`, and `Fatal` functions with `nil` context parameters so that they use `GetCallingFunctionName` instead.